+++ src/lib.rs
@@ -44,6 +44,7 @@ mod renderer;
 mod ui;
 mod widget;
 
+use self::event::CustomEvent;
 use self::placeholder::Placeholder;
 
 pub use self::event::Event;